+++ CHANGELOG.md
@@ -25,6 +25,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Constant propagation is now a proper must-analysis, if a variable is undefined
   in some path then it will be considered as non-constant
 - Dataflow: Only consider reachable nodes, which prevents some FPs/FNs
+- Timing output handles errors and reports profiling times
 - semgrep-core will log a warning when a worker process is consuming above 400 MiB
   of memory, or reached 80% of the specified memory limit, whatever happens first.
   This is meant to help diagnosing OOM-related crashes.